Add module to lock down kernel module loading
Modules can still be loaded during initrd time, but will
be disallowed after root has been pivoted. This significantly
increases the difficulty to inject a non-persistant root kit
during run time.

It must be explicitly included using -a, and will be enforced
unless rd.lockdown=0 is passed, which implies that GRUB must
be protected as well.